Initial Commit.
# FTC 16760 and 28147

This repository contains the public FTC SDK for the 2024-2025 season Into the Deep,
along with the RoadRunner source and a localizer for RoadRunner using GoBilda Pinpoint developed by J5155 and FormulaCarbon

This branch contains only this code. Official code should be held in branches with the team name.

For example, the 16760 branch will be called FTC-16760 and the 28147 branch will be called FTC-28147.
If you would like to branch off of your branch, call it FTC-teamnumber-branchname.

These actions will aid in making a neat codebase for our teams.